refactor: extract and move AppHeaderUnderline, FixedHeader and StyledHeader
