+++ src/components/AppComponents/LayoutFixedHeader/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./LayoutFixedHeader";
