+++ src/components/AppComponents/StyledHeader/index.ts
@@ -1,0 +1,2 @@
+export { StyledHeader, baseHeaderStyles } from "./StyledHeader";
+export type { HeaderProps } from "./StyledHeader";
