--- src/components/Header/index.ts
@@ -1,3 +1,0 @@
-export { default } from "./FixedHeader";
-export { StyledHeader } from "./StyledHeader";
-export { HeaderUnderline } from "./HeaderUnderline";
